sales_predictor_app/requirements.txt
streamlit
PyPDF2
pytesseract
pdf2image
Pillow
